Support automerge for CODEOWNERS PRs
This requires intentionally circumventing branch protection rules.
Since this is a metadata-only change, this is likely to be deemed
an acceptable tradeoff.

NOTE: branch protection rules are complex, so you are advised to use
DEBUG=true with RUNMODE=merge_pulls, as this will preserve all branch
protection rules to disk prior to modifying them.